+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 2.2.0-SNAPSHOT
+
+1. Send hidden claim messages for POAPs
+2. Fix distribution of failed attendees
+3. Display support error message on failure
+
 ## 2.1.0-RELEASE (2021-11-23)
 
 1. Add Privacy Policy and Terms & Conditions to /docs